feat: migrate CLI tests to tests/e2e/ directory (Phase 3)
Create tests/e2e/ directory and migrate 9 CLI test files from
tests/integration/, split test_e2e_phase6.py into 4 focused files,
rename test_e2e_scenarios.py to test_scenario_flows.py, and delete
originals. All 1344 tests pass. Tasks T008-T026 complete.
