+++ tests/e2e/__init__.py
@@ -1,0 +1,1 @@
+"""E2E (end-to-end) tests for fdsx CLI."""
