+++ tests/e2e/cli_test_utils.py
