+++ tests/e2e/conftest.py
@@ -1,0 +1,1 @@
+# No additional fixtures needed beyond tests/conftest.py (autouse _isolate_cwd).
