+++ tests/e2e/test_cli_flow_types.py
@@ -1,4 +1,4 @@
-from tests.integration.cli_test_utils import fixture_path, run_fdsx
+from tests.e2e.cli_test_utils import fixture_path, run_fdsx
 
 
 class TestCLIE2EPhase2:
