+++ tests/e2e/test_cli_signal_handling.py
